Basic GUI
this program consist of basic Gui for the program , it has three button in the main frame .One that will change frame to Sign up frame ,One that will change frame to Log in and last one that will quite the program. There are different input fields for user in sign up form.